backend/requirements.txt
# --- Security-patched versions (see SECURITY.md) ---
fastapi>=0.115.0
uvicorn[standard]>=0.27.0
# python-multipart >=0.0.22 fixes CVE-2026-24486 (path traversal) and CVE-2024-53981
python-multipart>=0.0.22
pydantic>=2.9.0
pydantic-settings>=2.5.0
structlog==25.5.0
orjson==3.11.5
webdavclient3==3.14.6
# jinja2 >=3.1.6 fixes CVE-2024-22195, CVE-2024-34064, CVE-2024-56326, CVE-2024-56201, CVE-2025-27516
jinja2>=3.1.6
aiosmtplib==3.0.1
sqlalchemy==2.0.25
alembic==1.13.1
# python-jose removed: PYSEC-2024-232/233 (DoS + algorithm confusion).
# PyJWT is the actively maintained replacement used for short-lived upload session tokens.
PyJWT>=2.8.0
passlib[bcrypt]==1.7.4
pytest==7.4.4
pytest-asyncio==0.23.3
httpx>=0.27.0
python-dotenv==1.0.0
email-validator==2.1.0
# Rate limiting (OWASP A04 – Insecure Design)
slowapi>=0.1.9
# Magic-bytes file type verification (OWASP A01 / CWE-434 – Unrestricted File Upload)
filetype>=1.2.0
